+++ Solutions/CipherCloud/Data Connectors/CipherCloudConnector/CipherCloud_API_FunctionApp.json
@@ -1,7 +1,7 @@
 {
-    "id": "GCPIAMDataConnector",
-    "title": "Google Cloud Platform IAM",
-    "publisher": "Google",
+    "id": "CipherCloudDataConnector",
+    "title": "Cipher Cloud",
+    "publisher": "Lookout",
     "descriptionMarkdown": "The Google Cloud Platform Identity and Access Management (IAM) data connector provides the capability to ingest [GCP IAM logs](https://cloud.google.com/iam/docs/audit-logging) into Microsoft Sentinel using the GCP Logging API. Refer to [GCP Logging API documentation](https://cloud.google.com/logging/docs/api) for more information.",
     "additionalRequirementBanner": ">This data connector depends on a parser based on a Kusto Function to work as expected [**GCP_IAM**](https://aka.ms/sentinel-GCPIAMDataConnector-parser) which is deployed with the Microsoft Sentinel Solution.",
     "graphQueries": [
