+++ Solutions/CipherCloud/Data Connectors/CipherCloudConnector/azuredeploy_Cipher_Cloud_FunctionApp.json
@@ -29,7 +29,11 @@
         "WorkspaceKey": {
             "type": "securestring",
             "defaultValue": ""
-        }
+        },
+		 "Baseurl": {
+            "type": "string",
+            "defaultValue": ""
+        }, 
     },
     "variables": {
         "FunctionName": "[concat(toLower(parameters('FunctionName')), uniqueString(resourceGroup().id))]",
@@ -157,11 +161,12 @@
                         "APPLICATIONINSIGHTS_CONNECTION_STRING": "[reference(resourceId('microsoft.insights/components', variables('FunctionName')), '2015-05-01').ConnectionString]",
                         "AzureWebJobsStorage": "[concat('DefaultEndpointsProtocol=https;AccountName=', toLower(variables('FunctionName')),';AccountKey=',listKeys(resourceId('Microsoft.Storage/storageAccounts', toLower(variables('FunctionName'))), '2019-06-01').keys[0].value, ';EndpointSuffix=',toLower(variables('StorageSuffix')))]",
                         "LookoutClientId": "[parameters('LookoutClientId')]",
-				"LookoutApiSecret":"[parameters('LookoutApiSecret')]",                        
-                        "WORKSPACE_ID": "[parameters('WorkspaceID')]",
-                        "SHARED_KEY": "[parameters('WorkspaceKey')]",
+				        "LookoutApiSecret":"[parameters('LookoutApiSecret')]",                        
+                        "WorkspaceID": "[parameters('WorkspaceID')]",
+                        "WorkspaceKey": "[parameters('WorkspaceKey')]",
                         "logAnalyticsUri": "[variables('LogAnaltyicsUri')]",
-                        "WEBSITE_RUN_FROM_PACKAGE": "https://aka.ms/sentinel-GCPIAMDataConnector-functionapp"
+						"Baseurl":"[parameters('Baseurl')]",
+                        "WEBSITE_RUN_FROM_PACKAGE": "https://github.com/Azure/Azure-Sentinel/blob/vetadika_CipherCloud/Solutions/CipherCloud/Data%20Connectors/CipherCloudConnector/CipherCloudConnector.zip?raw=true"
                     }
                 }
             ]
