+++ Solutions/CipherCloud/Data Connectors/CipherCloudConnector/requirements.txt
@@ -2,7 +2,9 @@
 # The Python Worker is managed by Azure Functions platform
 # Manually managing azure-functions-worker may cause unexpected issues
 
-azure-storage-file-share==12.3.0
-azure-functions
 pyjwt
-requests
+azure-functions
+azure-storage-file-share==12.3.0
+python-dateutil==2.8.1
+azure_core==2.0.12
+requests==2.25.1
