+++ Solutions/CipherCloud/Parsers/CipherCloudAnomalies.txt
@@ -1,0 +1,18 @@
+// Usage Instruction :
+// Paste below query in log analytics, click on Save button and select as Function from drop down by specifying function name and alias as CipherCloudEvents.
+// Function usually takes 10-15 minutes to activate. You can then use function alias from any other queries (e.g. CipherCloudAnamolies | take 10).
+// Reference : Using functions in Azure monitor log queries : https://docs.microsoft.com/azure/azure-monitor/log-query/functions
+Lookoutlogs_CL
+|extend timeStamp = timeStamp_t
+| extend eventType = eventType_s
+|extend  actionType = actionType_s
+|extend  anomalyType = anomalyType_s
+|extend  appName = appName_s
+|extend userEmail = userEmail_s
+|extend  anomalyName = anomalyName_s
+|extend  currentCity = currentCity_s
+|extend currentTimestamp = currentTimestamp_t
+|extend previousTimestamp = previousTimestamp_t
+| extend  currentEventId = currentEventId_g
+|extend  previousEventId = previousEventId_g
+| where eventType == 'Anomaly'
